+++ src-docs/src/views/link/link_example.js
@@ -65,10 +65,11 @@ export const LinkExample = {
       ],
       text: (
         <p>
-          When an <EuiCode>EuiLink</EuiCode> is passed an onClick method, and is
-          not passed an <EuiCode>href</EuiCode>, it can optionally be set to
+          When an <EuiCode>EuiLink</EuiCode> is passed an{' '}
+          <EuiCode>onClick</EuiCode> method, and is not passed an{' '}
+          <EuiCode>href</EuiCode>, it can optionally be set to
           <EuiCode>disabled</EuiCode> which disables the click behavior, and
-          hides the link.
+          removes the link styling.
         </p>
       ),
       props: { EuiLink },